Resolve version requirements when looking for unneeded packages
Fixes #1335
